chore: cconsolidated alignment scoffre matrices into align package
