+++ align/align.go
@@ -63,6 +63,13 @@ var HumanChimpTwoScoreMatrix = [][]int64{
 	{-208, -196, -196, -208, -202},
 }
 
+var HumanChimpTwoScoreMatrixNoGap = [][]int64{
+	{90, -330, -236, -356},
+	{-330, 100, -318, -236},
+	{-236, -318, 100, -330},
+	{-356, -236, -330, 90},
+}
+
 /*
 var StrictScoreMatrix = [][]int64{
         {91, -114, -31, -123, -44},
