+++ genomeGraph/localAlignment_test.go
@@ -4,6 +4,7 @@ import (
 	"log"
 	"testing"
 
+	"github.com/vertgenlab/gonomics/align"
 	"github.com/vertgenlab/gonomics/cigar"
 	"github.com/vertgenlab/gonomics/dna"
 )
@@ -15,7 +16,7 @@ func TestRightLocal(t *testing.T) {
 	var seqOneB = dna.StringToBases("ATTTTTTTTTTTTTTTTAGC")
 	m, trace := swMatrixSetup(10000)
 
-	score, alignmentPath, refStart, refEnd, queryStart, queryEnd := RightLocal(seqOneA, seqOneB, HumanChimpTwoScoreMatrix, -600, m, trace)
+	score, alignmentPath, refStart, refEnd, queryStart, queryEnd := RightLocal(seqOneA, seqOneB, align.HumanChimpTwoScoreMatrix, -600, m, trace)
 
 	log.Printf("score=%d, alignment=%s, refStart=%d, refEnd=%d, queryStart=%d, queryEnd=%d\n", score, cigar.ToString(alignmentPath), refStart, refEnd, queryStart, queryEnd)
 }
@@ -25,7 +26,7 @@ func TestLeftLocal(t *testing.T) {
 	var seqOneB = dna.StringToBases("CAGGGGGTGGGGGGGG")
 	m, trace := swMatrixSetup(10000)
 
-	score, alignmentPath, refStart, refEnd, queryStart, queryEnd := LeftLocal(seqOneA, seqOneB, HumanChimpTwoScoreMatrix, -600, m, trace)
+	score, alignmentPath, refStart, refEnd, queryStart, queryEnd := LeftLocal(seqOneA, seqOneB, align.HumanChimpTwoScoreMatrix, -600, m, trace)
 
 	log.Printf("score=%d, alignment=%s, refStart=%d, refEnd=%d, queryStart=%d, queryEnd=%d\n", score, cigar.ToString(alignmentPath), refStart, refEnd, queryStart, queryEnd)
 }
