+++ genomeGraph/view.go
@@ -2,17 +2,19 @@ package genomeGraph
 
 import (
 	"bytes"
+	"strings"
+
+	"github.com/vertgenlab/gonomics/align"
 	"github.com/vertgenlab/gonomics/cigar"
 	"github.com/vertgenlab/gonomics/dna"
 	"github.com/vertgenlab/gonomics/numbers/parse"
 	"github.com/vertgenlab/gonomics/sam"
-	"strings"
 )
 
 func calcExtension(seq []dna.Base) int64 {
 	var maxScore int64 = 0
 	for bases := 0; bases < len(seq); bases++ {
-		maxScore += HumanChimpTwoScoreMatrix[seq[bases]][seq[bases]]
+		maxScore += align.HumanChimpTwoScoreMatrix[seq[bases]][seq[bases]]
 	}
 	return maxScore/600 + int64(len(seq))
 }
